backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
pydantic==2.5.0
python-dotenv==1.0.0
transformers==4.35.2
torch==2.1.1
pdfplumber==0.10.3
python-docx==1.1.0
spacy==3.7.2
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.1.4
celery==5.3.4
redis==5.0.1
reportlab==4.0.7
Pillow==10.1.0
aiofiles==23.2.1